Starting on adding items to the trolley.
{
   Viewing menu works
   Check in system works
    Log in system works.
    Register system works
    Save user session works
}

+++ app/src/main/java/com/uclan/ashleymorris/goeat/Classes/Item.java
@@ -11,14 +11,14 @@ package com.uclan.ashleymorris.goeat.Classes;
 public class Item {
 
     private String name;
-    private Float price;
+    private double price;
     private String description;
 
     public Item(){
 
     }
 
-    public Item(String name, Float price, String description) {
+    public Item(String name, Double price, String description) {
         this.name = name;
         this.price = price;
         this.description = description;
@@ -28,7 +28,7 @@ public class Item {
         return name;
     }
 
-    public Float getPrice() {
+    public double getPrice() {
         return price;
     }
